preparation/mc4/requirements.txt
ftfy
google-cloud-storage
tqdm
bettermap
more_itertools

tensorflow
protobuf<3.21.0